feat: solve task 04-copy-directory

+++ 04-copy-directory/index.js
@@ -2,33 +2,36 @@ const path = require('path');
 const fs = require('fs');
 const { rm ,copyFile, mkdir } = require('fs/promises');
 
-const copyDir = async(src) => {
+const copyDir = async(src, input = __dirname, output = __dirname) => {
+  const folder = path.join(input, `${src}`);
+  const copyFolder = path.join(output, input === __dirname ? `${src}-copy` : `${src}`);
   try {
-    const folder = path.join(__dirname, `${src}`);
-    const copyFolder = path.join(__dirname, `${src}-copy`);
-    try {
-      await fs.promises.rm(copyFolder, { recursive: true });
-      console.log(`Папка ${src}-copy реинициализирована`);
-    }
-    catch (err) {
-      console.log(`Папка ${src}-copy создана`);
-    }
-    finally {
-      await mkdir(copyFolder, { recursive: true });
-    }
-  
-    fs.readdir(folder, (err, arr) => {
-      if (err) throw err;
-      else {
-        arr.forEach(async(data) => {
-          await copyFile(path.join(folder, data), path.join(copyFolder, data));
-        })
-      }
-    });
+    await rm(copyFolder, { recursive: true });
+    console.log(`Папка ${src} реинициализирована`);
   }
   catch (err) {
-    throw err;
+    console.log(`Папка ${src} создана`);
   }
-}
+  finally {
+    await mkdir(copyFolder, { recursive: true });
+  }
+
+  fs.readdir(folder, (err, arr) => {
+    if (err) throw err;
+    else {
+      arr.forEach((data) => {
+        fs.stat(path.join(folder, data), (err, stats) => {
+          if (err) throw err;
+          else if (stats.isFile()) {
+            copyFile(path.join(folder, data), path.join(copyFolder, data));
+          }
+          else if(stats.isDirectory()) {
+            copyDir(data, folder, copyFolder);
+          }
+        });
+      });
+    }
+  });
+};
 
 copyDir('files');
